+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `REPEATABLESET` predicate type to `conditionalLogicService.evaluateConditionalPredicate`
+
 ## [4.1.1] - 2023-10-15
 
 ### Added
 
 - CI job for building docs
-- `REPEATABLESET` predicate type to `conditionalLogicService.evaluateConditionalPredicate`
 
 ## [4.1.0] - 2023-08-29